+++ src/app/components/ErrorHandler/ErrorHandler.tsx
@@ -1,5 +1,5 @@
 import { ERROR_HANDLING_STRATEGY } from '@constants/constants';
-import { returnErrorHandlersNames, toHtmlSafeId } from '@utils/shared';
+import { toHtmlSafeId } from '@utils/shared';
 import { t } from 'i18next';
 import React from 'react';
 
@@ -21,6 +21,42 @@ type ErrorHandlerProps = {
     checked: boolean,
     event: React.FormEvent<HTMLInputElement>
   ) => void;
+};
+
+type ErrorHandlerObj = {
+  errorHandler: string;
+  description: string;
+};
+/**
+ * This method returns the error handler object based on the error handler strategy string
+ * @param errorHandler
+ * @returns ErrorHandlerObj
+ */
+export const returnErrorHandlersNames = (
+  errorHandler: string
+): ErrorHandlerObj => {
+  switch (errorHandler) {
+    case ERROR_HANDLING_STRATEGY.Log:
+      return {
+        errorHandler: t('ignore'),
+        description: t('ignoreDescription'),
+      };
+    case ERROR_HANDLING_STRATEGY.DeadLetterQueue:
+      return {
+        errorHandler: t('deadLetterQueue'),
+        description: t('deadLetterQueueDescription'),
+      };
+    case ERROR_HANDLING_STRATEGY.Stop:
+      return {
+        errorHandler: t('stop'),
+        description: t('stopDescription'),
+      };
+    default:
+      return {
+        errorHandler: errorHandler,
+        description: '',
+      };
+  }
 };
 
 export const ErrorHandler: React.FunctionComponent<ErrorHandlerProps> = ({
